+++ devo/common/__init__.py
@@ -1,10 +1,14 @@
-from .dates.dateparser import parse, parse_string, parse_expression, \
-    default_from, default_to
-from .dates.dateoperations import month, week, day, hour, minute, second, now, \
-    now_without_ms, today, yesterday, parse_functions
-from .dates.dateutils import to_millis, trunc_time, trunc_time_minute, \
-    test_date_format, get_timestamp
+from .dates.dateparser import (
+    parse, parse_string, parse_expression,
+    default_from, default_to)
+from .dates.dateoperations import (
+    month, week, day, hour, minute, second, now,
+    now_without_ms, today, yesterday, parse_functions)
+from .dates.dateutils import (
+    to_millis, trunc_time, trunc_time_minute,
+    test_date_format, get_timestamp)
 from .generic.configuration import Configuration
-from .logging.log import get_log, set_formatter, get_rotating_file_handler, \
-    get_stream_handler
+from .logging.log import (
+    get_log, set_formatter, get_rotating_file_handler,
+    get_stream_handler)
 from .loadenv import load_env_file